better constants, merge cpuidle into cpu
