+++ src/common/mod.rs
@@ -9,6 +9,23 @@ pub mod bpf;
 pub const VERSION: &str = env!("CARGO_PKG_VERSION");
 pub const NAME: &str = env!("CARGO_PKG_NAME");
 
+pub const SECOND: u64 = 1_000 * MILLISECOND;
+pub const MILLISECOND: u64 = 1_000 * MICROSECOND;
+pub const MICROSECOND: u64 = 1_000 * NANOSECOND;
+pub const NANOSECOND: u64 = 1;
+
+pub const TERABIT: u64 = 1_000 * GIGABIT;
+pub const GIGABIT: u64 = 1_000 * MEGABIT;
+pub const MEGABIT: u64 = 1_000 * KILOBIT;
+pub const KILOBIT: u64 = 1_000 * BIT;
+pub const BIT: u64 = 1;
+
+pub const TEBIBYTE: u64 = 1_024 * GIBIBYTE;
+pub const GIBIBYTE: u64 = 1_024 * MEBIBYTE;
+pub const MEBIBYTE: u64 = 1_024 * KIBIBYTE;
+pub const KIBIBYTE: u64 = 1_024 * BYTE;
+pub const BYTE: u64 = 1;
+
 /// helper function to discover the number of hardware threads
 pub fn hardware_threads() -> Result<u64, ()> {
     let path = "/sys/devices/system/cpu/present";
