+++ src/samplers/disk/mod.rs
@@ -2,6 +2,7 @@
 // Licensed under the Apache License, Version 2.0
 // http://www.apache.org/licenses/LICENSE-2.0
 
+use crate::common::*;
 use crate::common::bpf::*;
 use crate::config::{Config, SamplerConfig};
 use crate::samplers::Common;
@@ -151,7 +152,7 @@ impl Sampler for Disk {
                                     self.metrics().record_distribution(
                                         statistic,
                                         time,
-                                        value * 1000,
+                                        value * MICROSECOND,
                                         count,
                                     );
                                 }
@@ -178,9 +179,9 @@ impl Sampler for Disk {
         };
 
         let max = if statistic.ebpf_table().is_some() {
-            1_000_000
+            SECOND
         } else {
-            1_000_000_000_000
+            TEBIBYTE
         };
 
         Some(Summary::histogram(
